+++ test/integration/CMakeLists.txt
@@ -48,6 +48,7 @@ add_executable(
     test/snippets/multi_function.cpp
     test/snippets/metadata.cpp
     test/snippets/charsets.cpp
+    test/snippets/time_types.cpp
 )
 target_include_directories(
     boost_mysql_integrationtests
